1. 完成修改收费功能 2. 移植收费记录功能 3. 完成修改全站添加功能

+++ src/main/java/com/keymanager/ckadmin/service/CustomerService.java
@@ -36,5 +36,5 @@ public interface CustomerService extends IService<Customer> {
 
     List<Customer> getActiveCustomerSimpleInfo(CustomerCriteria customerCriteria);
 
-    Customer getCustomer(int customerUuid);
+    Customer getCustomer(Long customerUuid);
 }
